% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{exh.train.cv}
\alias{exh.train.cv}
\title{Cross-validation of a panel}
\usage{
exh.train.cv(data, predictors, response, fixed.predictors = NULL,
  id = NULL, levels = base::levels(as.factor(data[[response]])),
  na.rm = FALSE, nreps = 10, k = 10, stratified = TRUE,
  verbose = FALSE, working.dir = NULL, java.keep.files = TRUE, ...)
}
\arguments{
\item{data}{the dataset}

\item{predictors}{the name of the columns to be tested as predictors in the panel. May or may not be used.}

\item{response}{the binary response column name}

\item{fixed.predictors}{predictors to force into the panel}

\item{id}{the name of a column storing unique IDs for the observations}

\item{levels}{the values of \code{response} to use as negative and positive data}

\item{na.rm}{whether to omit rows with \code{NA} observations}

\item{nreps}{number of cross-validation repeats}

\item{k}{number of folds of the cross-validation (use \code{k = nrow(data)}) for leave-one-out CV}

\item{stratified}{whether to keep the balance of positive/negative samples in all the CV folds}

\item{verbose}{enables additional output for debugging}

\item{working.dir, java.keep.files}{same as for \code{\link{exh.train}}}

\item{...}{further arguments passed to \code{\link{exh.train}}. \code{constrain.on}, \code{min.constr}, \code{panels.of.num} or \code{limit.java.threads} are probably the most useful ones.}
}
\description{
Calls exh.train in nreps k-fold CV round to evaluate performance
}
\examples{
data(aSAH, package="pROC")
exh.train.cv(aSAH, c("age", "s100b", "ndka"), "outcome")
}
