% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{[[.exhlist}
\alias{[[.exhlist}
\title{Extraction of an \code{exhlist}}
\usage{
\method{[[}{exhlist}(x, i)
}
\arguments{
\item{x}{the \code{exhlist} from \code{\link{exh.train}}}

\item{i}{index to extract}
}
\description{
Extraction of an \code{exhlist}
}
