% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internals.R, R/cv.R
\name{Internals}
\alias{Internals}
\alias{cv.indices}
\title{Internal functions not intended for general use. These functions may be changed or removed at any time without notice.}
\usage{
cv.indices(n, k, stratified = TRUE, response, levels)
}
\arguments{
\item{n}{number of observations}

\item{k}{number of folds}

\item{stratified}{if CV must be stratified (similar number of observations of each group in each fold) or not. Defaults to \code{TRUE}.}

\item{response, levels}{as for \code{\link{exh.train.cv}}}
}
\description{
Internal functions not intended for general use. These functions may be changed or removed at any time without notice.
}
