% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr.panels.R
\name{nr.panels}
\alias{nr.panels}
\alias{nr.panels.exhlist}
\alias{nr.panels.exhcv}
\alias{nr.panels.exhcvlist}
\title{Counts the number of panels found}
\usage{
nr.panels(x)

\method{nr.panels}{exhlist}(x)

\method{nr.panels}{exhcv}(x)

\method{nr.panels}{exhcvlist}(x)
}
\arguments{
\item{x}{the panels as an \code{exhlist}, single \code{exhcv} or \code{exhcvlist}}
}
\description{
Counts the number of panels found
}
