% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv.diag}
\alias{plot.cv.diag}
\title{Plot diagnosis on cross validation}
\usage{
\method{plot}{cv.diag}(x, data, training.panels, ...)
}
\arguments{
\item{x}{the cross-validation of a panel (\code{\link{exh.train.cv}})}

\item{data}{some data for the plot}

\item{training.panels}{not sure about that one}

\item{...}{ignored}
}
\description{
Plot diagnosis on cross validation
}
