% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.exh}
\alias{plot.exh}
\title{Plot the MDS of a panel}
\usage{
\method{plot}{exh}(
  x,
  newdata = NULL,
  xlab = "First MDS dimension",
  ylab = "Second MDS dimension",
  main = "MDS on PRIM",
  ask = dev.interactive(),
  ...
)
}
\arguments{
\item{x}{an train panel}

\item{newdata}{the data to plot}

\item{xlab, ylab, main}{labels for the plot}

\item{ask}{ask to create new pages of the plot}

\item{...}{further arguments for \code{\link{plot}}}
}
\description{
Plot the MDS of a panel
}
