% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.thr.stability.R
\name{plot.thr.stability}
\alias{plot.thr.stability}
\alias{plot.thr.stability.exhlist}
\alias{plot.thr.stability.exhcv}
\alias{plot.thr.stability.exhcvlist}
\title{Plots the stability of thresholds over cross-validation}
\usage{
plot.thr.stability(...)

\method{plot.thr.stability}{exhlist}(x, data = NULL, found.mols = NULL,
  add = FALSE, plot = TRUE, ...)

\method{plot.thr.stability}{exhcv}(x, data, found.mols = NULL,
  add = FALSE, plot = TRUE, ...)

\method{plot.thr.stability}{exhcvlist}(x, data, found.mols = NULL,
  add = FALSE, plot = TRUE, ...)
}
\arguments{
\item{...}{additional arguments to \code{\link{plot}}}

\item{x}{the panels as an \code{exhlist}, single \code{exhcv} or \code{exhcvlist}}

\item{data}{data to be plotted}

\item{found.mols}{the vector of molecules that were found. Automatically extracted from \code{x} unless provided}

\item{add}{whether to just add to the plot}

\item{plot}{whether to actually plot that part}
}
\description{
Plots the stability of thresholds over cross-validation
}
