% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.exh}
\alias{predict.exh}
\alias{predict}
\alias{predict.exhlist}
\alias{predict.exhcv}
\alias{predict.exhcvlist}
\title{Model predictions}
\usage{
\method{predict}{exh}(object, newdata = NULL, center = TRUE, ...)

\method{predict}{exhlist}(object, newdata = NULL, statistics = mean, ...)

\method{predict}{exhcv}(object, ...)

\method{predict}{exhcvlist}(object, statistics = mean, ...)
}
\arguments{
\item{object}{a model: \code{exh}, \code{exhlist}, \code{exhcv} or \code{exhcvlist}}

\item{newdata}{some data to make the predictions}

\item{center}{should the predicted values be centered around 0? If TRUE, returned values will be comprized between -1 and +1, with mean 0. Useful when multiple panels have a different \code{min.nr} and averaging them doesn't can't be done directly}

\item{...}{passing arguments to and from other methods}

\item{statistics}{how to average}
}
\description{
Model predictions
}
\examples{
data(aSAH, package = "pROC")
panels <- exh.train(aSAH, c("age", "s100b", "ndka"), "outcome", progress=FALSE)
predict(panels, aSAH)
}
