% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.exhlist}
\alias{print.exhlist}
\alias{print.exhcvlist}
\alias{print.exhcv}
\alias{print.exh}
\title{Prints a panel}
\usage{
\method{print}{exhlist}(x, file = "", append = FALSE, ...)

\method{print}{exhcvlist}(x, file = "", append = FALSE, ...)

\method{print}{exhcv}(x, file = "", append = FALSE, ...)

\method{print}{exh}(x, file = "", append = FALSE, ...)
}
\arguments{
\item{x}{the panel: \code{exh}, \code{exhlist}, \code{exhcv} or \code{exhcvlist}}

\item{file}{if the empty string \dQuote{}, prints on the R output}

\item{append}{whether to erase the file before printing. See \code{\link{cat}}}

\item{...}{additional arguments passed from and to other methods}
}
\description{
Prints a panel
}
