% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{read.java.output}
\alias{read.java.output}
\title{Reads the output of the java PanelomiX processing. If you plan to re-read a panel later, make sure you run \code{\link{exh.train}} with \code{java.keep.files=TRUE}.}
\usage{
read.java.output(conf.file = "conf", result.file = NA)
}
\arguments{
\item{conf.file}{the name of the configuration file to read data from}

\item{result.file}{can be used to restrict which files to read exactly}
}
\description{
Reads the output of the java PanelomiX processing. If you plan to re-read a panel later, make sure you run \code{\link{exh.train}} with \code{java.keep.files=TRUE}.
}
