% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.mol.stability.R
\name{table.mol.stability}
\alias{table.mol.stability}
\alias{table.mol.stability.exhcvlist}
\alias{table.mol.stability.exhcv}
\alias{table.mol.stability.exhlist}
\title{Frequency of the choice of molecules}
\usage{
table.mol.stability(object, ...)

\method{table.mol.stability}{exhcvlist}(object, ...)

\method{table.mol.stability}{exhcv}(object, ...)

\method{table.mol.stability}{exhlist}(object, ...)
}
\arguments{
\item{object}{the panel model}

\item{...}{additional arguments to and from other methods}
}
\description{
Frequency of the choice of molecules
}
\examples{
data(aSAH, package = "pROC")
cv <- exh.train.cv(aSAH, c("age", "s100b", "ndka"), "outcome", progress=FALSE)
table.mol.stability(cv)
}
