% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.perf.R
\name{train.perf}
\alias{train.perf}
\alias{train.perf.exh}
\alias{train.perf.exhlist}
\alias{train.perf.exhcv}
\alias{train.perf.exhcvlist}
\title{Performance on the training data}
\usage{
train.perf(...)

\method{train.perf}{exh}(object, ...)

\method{train.perf}{exhlist}(object, ...)

\method{train.perf}{exhcv}(object, ...)

\method{train.perf}{exhcvlist}(object, ...)
}
\arguments{
\item{...}{additional arguments to and from other methods}

\item{object}{the model}
}
\description{
Calculates the fit on the training data. See \code{\link{test.perf}} to test the fit on a new data set
}
