% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.rjava.R
\name{exh.train}
\alias{exh.train}
\title{Train a panel with the new java interface}
\usage{
exh.train(data, predictors, response, fixed.predictors = NULL,
  id = NULL, constrain.on = c("specificity", "sensitivity",
  "accuracy"), min.constr = 0.95,
  levels = base::levels(as.factor(data[[response]])),
  panels.of.num = 1:length(predictors), test.thresholds = NA,
  directions = NULL, filter.number.thresholds = NA,
  filter.randomForest = FALSE, verbose = TRUE,
  multiple.panels = c("all", "random", "first"), na.rm = c("all",
  "local"), rand.response = FALSE, limit.java.threads = NA,
  working.dir = NULL, progress = TRUE, ...)
}
\arguments{
\item{data}{the dataset}

\item{predictors}{the name of the columns to be tested as predictors in the panel. May or may not be used.}

\item{response}{the binary response column name}

\item{fixed.predictors}{predictors to force into the panel}

\item{id}{a column with sample ids}

\item{constrain.on, min.constr}{objective constrains. Consider only panel with \code{min.nr} \dQuote{specificity} or \dQuote{sensitivity}, or maximize global \dQuote{accuracy}}

\item{levels}{the values of \code{response} to use as negative and positive data}

\item{panels.of.num}{a vector of integer, defining the acceptable number of markers included in the panel}

\item{test.thresholds}{the thresholds to test. Ignored if filter.randomForest is a list}

\item{directions}{a named list (after the \code{predictors} and \code{fixed.predictors}) of directions as described in \code{\link[pROC]{roc}}}

\item{filter.number.thresholds}{if \code{TRUE}, filter the \code{test.thresholds} to speed up the search. Ignored if filter.randomForest is a list. Can be the character \dQuote{2x2} or a numeric (length 1)}

\item{filter.randomForest}{a \code{\link{list}} with 2 elements: \code{molecules} (number of markers to keep) \code{thresholds} (number of thresholds to keep per marker)}

\item{verbose}{enables additional output for debugging}

\item{multiple.panels}{one of \dQuote{all}, \dQuote{random} or \dQuote{first} when multiple panels are equivalent, which one(s) to report}

\item{na.rm}{one of \dQuote{all} or \dQuote{local}}

\item{rand.response}{randomize the response.}

\item{limit.java.threads}{how many threads to use. If \code{NA} (default), java will define itself how many threads to use. If an integer is given, will force to this number}

\item{working.dir}{a (preferably empty) directory where the files required by java will be placed. If \code{\link{missing}}, \code{NULL} \code{NA} or an empty string, a temporary folder will be created and removed afterwards, otherwise the given string will be used and files will be kept in place. For example \code{\link{tempdir}()}.}

\item{progress}{whether to display a graphical progress bar}

\item{...}{additional arguments passed from and to other methods. In particular, further arguments for to \code{\link[pROC]{roc}}.}
}
\description{
Train a panel with the new java interface
}
\examples{
data(aSAH, package = "pROC")
exh.train(aSAH, c("age", "s100b", "ndka"), "outcome")
}
