% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.nr.stability.R
\name{table.nr.stability}
\alias{table.nr.stability}
\alias{table.nr.stability.exhcvlist}
\alias{table.nr.stability.exhcv}
\alias{table.nr.stability.exhlist}
\title{Frequency of the number of molecules included}
\usage{
table.nr.stability(object, ...)

\method{table.nr.stability}{exhcvlist}(object, ...)

\method{table.nr.stability}{exhcv}(object, ...)

\method{table.nr.stability}{exhlist}(object, ...)
}
\arguments{
\item{object}{the panel model}

\item{...}{additional arguments to and from other methods}
}
\description{
Frequency of the number of molecules included
}
\examples{
data(aSAH, package = "pROC")
cv <- exh.train.cv(aSAH, c("age", "s100b", "ndka"), "outcome", progress=FALSE)
table.nr.stability(cv)
}
